stretchedfixedmode: rename plugin to 'stretched mode'
