+++ runelite-client/src/main/java/net/runelite/client/plugins/stretchedmode/TranslateMouseWheelListener.java
@@ -23,68 +23,32 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.stretchedfixedmode;
+package net.runelite.client.plugins.stretchedmode;
 
 import java.awt.Component;
 import java.awt.Dimension;
-import java.awt.event.MouseEvent;
+import java.awt.event.MouseWheelEvent;
 import javax.inject.Inject;
 import net.runelite.api.Client;
-import net.runelite.client.input.MouseListener;
+import net.runelite.client.input.MouseWheelListener;
 
-public class TranslateMouseListener implements MouseListener
+public class TranslateMouseWheelListener implements MouseWheelListener
 {
 	private final Client client;
 
 	@Inject
-	public TranslateMouseListener(Client client)
+	public TranslateMouseWheelListener(Client client)
 	{
 		this.client = client;
 	}
 
 	@Override
-	public MouseEvent mouseClicked(MouseEvent mouseEvent)
+	public MouseWheelEvent mouseWheelMoved(MouseWheelEvent event)
 	{
-		return translateEvent(mouseEvent);
+		return translateEvent(event);
 	}
 
-	@Override
-	public MouseEvent mousePressed(MouseEvent mouseEvent)
-	{
-		return translateEvent(mouseEvent);
-	}
-
-	@Override
-	public MouseEvent mouseReleased(MouseEvent mouseEvent)
-	{
-		return translateEvent(mouseEvent);
-	}
-
-	@Override
-	public MouseEvent mouseEntered(MouseEvent mouseEvent)
-	{
-		return translateEvent(mouseEvent);
-	}
-
-	@Override
-	public MouseEvent mouseExited(MouseEvent mouseEvent)
-	{
-		return translateEvent(mouseEvent);
-	}
-
-	@Override
-	public MouseEvent mouseDragged(MouseEvent mouseEvent)
-	{
-		return translateEvent(mouseEvent);
-	}
-
-	@Override
-	public MouseEvent mouseMoved(MouseEvent mouseEvent)
-	{
-		return translateEvent(mouseEvent);
-	}
-
-	private MouseEvent translateEvent(MouseEvent e)
+	private MouseWheelEvent translateEvent(MouseWheelEvent e)
 	{
 		Dimension stretchedDimensions = client.getStretchedDimensions();
 		Dimension realDimensions = client.getRealDimensions();
@@ -92,7 +56,7 @@ public class TranslateMouseListener implements MouseListener
 		int newX = (int) (e.getX() / (stretchedDimensions.width / realDimensions.getWidth()));
 		int newY = (int) (e.getY() / (stretchedDimensions.height / realDimensions.getHeight()));
 
-		return new MouseEvent((Component) e.getSource(), e.getID(), e.getWhen(), e.getModifiersEx(),
-				newX, newY, e.getClickCount(), e.isPopupTrigger(), e.getButton());
+		return new MouseWheelEvent((Component) e.getSource(), e.getID(), e.getWhen(), e.getModifiers(), newX, newY,
+				e.getClickCount(), e.isPopupTrigger(), e.getScrollType(), e.getScrollAmount(), e.getWheelRotation());
 	}
 }
